refactor: split feature flags into dedicated microcrate
