+++ copybook-feature-flags/README.md
@@ -1,0 +1,9 @@
+# copybook-feature-flags
+
+Standalone microcrate for feature-flag contracts and runtime controls used across the copybook-rs governance stack.
+
+## Provides
+
+- `Feature`, `FeatureCategory`, and `FeatureLifecycle`
+- `FeatureFlags`, `FeatureFlagsBuilder`, `FeatureFlagsHandle`
+- `all_features()` and environment-driven flag loading
